requirements.txt
pygame-ce
colorama
https://github.com/umutcanekinci/pygame-core/zipball/master
pyyaml